requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1
twilio>=8.10.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pillow>=10.1.0
requests>=2.31.0
aiofiles>=23.2.1
asyncpg>=0.29.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
httpx>=0.25.2

# Security & Rate Limiting
slowapi>=0.1.9
python-jose[cryptography]>=3.3.0

# Logging & Monitoring
structlog>=23.2.0
sentry-sdk[fastapi]>=1.38.0

# Development
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0 